Change tag comparison to be case insensitive
Tag comparison via Tag::equals (and Tag::hashCode) is currently
case-sensitive based on the tagName field.

In VolunteeRoll, tags are likely to be used for subjects (such as
Mathematics or English). Thus, it is neither necessary nor helpful for
tags to be case-sensitive.

Let's change the implementation of Tag::equals and Tag::hashCode to make
them case-insensitive.

Existing methods already use the equals() and (indirectly) hashCode()
methods to compare tags. Thus, changing these method implementations is
sufficient.

Regression testing was done to ensure there are no unintended side
effects to the rest of the project.