Add Javadoc comments for constructors
Constructors for FindByTagCommand and FindComment lack Javadoc header
comments.

Checkstyle raises an error due to the lack of these Javadoc comments.

Let's add Javadoc header comments for the constructors of these two
classes.

For discussion on why the missing Javadoc comment for FindComment::new
only raises a Checkstyle error now and not in AB3, refer to this forum
post:
https://github.com/nus-cs2103-AY2526S1/forum/issues/373