Undo renaming of Model and Logic method
The methods were renamed in this branch. They are widely used
throughout the codebase.

Renaming them may cause merge conflicts or break functionality when
updating this branch with changes to the 'master' branch.

Let's undo the renaming of these methods.

Handling this before merging 'master' onto this branch so that we
can easily redo naming with IDE's built-in refactoring tools.