Add context to start of UG
Some testers during PE-D and PE may be unfamiliar with such voluntary
tuition programmes in Singapore.

This may lead to confusion about the target audience of our product.

Let's add context to the start of the UG to help testers better
understand our target audience and the context of our app.

Context is added to the start of the UG which testers are likely to
read first or at least early on during testing, which will help
them better understand the rest of the documentation.